@@ -1,27 +1,39 @@
 {
-  "homepage": "http://brijkishorthakur.github.io/-Codepen-clone",
-  "name": "codepen-clone",
+  "name": "code_deck_coding_ide",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.3.1",
-    "@mui/material": "^5.3.1",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@codemirror/lang-cpp": "^6.0.2",
+    "@codemirror/lang-css": "^6.0.1",
+    "@codemirror/lang-html": "^6.4.0",
+    "@codemirror/lang-java": "^6.0.1",
+    "@codemirror/lang-javascript": "^6.1.1",
+    "@codemirror/lang-python": "^6.1.0",
+    "@codemirror/lang-xml": "^6.0.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "codemirror": "^5.65.1",
-    "react": "^17.0.2",
-    "react-codemirror2": "^7.2.1",
-    "react-dom": "^17.0.2",
-    "react-scripts": "5.0.0",
-    "web-vitals": "^2.1.4"
+    "@uiw/codemirror-theme-bespin": "^4.15.1",
+    "@uiw/codemirror-theme-dracula": "^4.15.1",
+    "@uiw/codemirror-theme-duotone": "^4.15.1",
+    "@uiw/codemirror-theme-github": "^4.15.1",
+    "@uiw/codemirror-theme-okaidia": "^4.15.1",
+    "@uiw/codemirror-theme-vscode": "^4.15.1",
+    "@uiw/codemirror-theme-xcode": "^4.15.1",
+    "@uiw/react-codemirror": "^4.15.1",
+    "buffer": "^6.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "react-select": "^5.6.1",
+    "uuid": "^9.0.0",
+    "web-vitals": "^2.1.4",
+    "react-icons": "4.7.1"
   },
   "scripts": {
-    "start": "react-scripts start",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
@@ -43,9 +55,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "gh-pages": "^6.1.1"
   }
 }
